Function: Retrieve Patient Entry Index Added
FHIR data contains a list of entries, one of which is the Patient resource type. The newly created method 'retrieve_patient_entry_index' will allow us to retrieve the index position of the Patient entry.
This will be useful when setting up the Patient Table in the database, allowing us to add patient data for each FHIR bundle.

It is noted that for the sample FHIR data, the patient entry is always the first entry in the list. As far as I can find, this is not a rule for all FHIR data, which is why I believe iterating through the list is the most reliable method.
A custom exception has been created 'PatientNotFoundError' which will be raised if the patient entry is not found.

Additional Changes:
- main has been replaced with utils.py, as this is a better description of the methods currently being created.