Function - Add Patient Entry Method created
In order to add the event resource type 'Patient' to the patient database, a new function - add_patient_entry has been created, which extracts various parts of the patient data, and adds it to the correct columns. The function returns the pid (int), a unique identifier for each patient which is intended to be referenced by the other event entries from the same patient.

The function also contains basic duplication handling, returning a minus value (-1) if the full_url value is a duplicate of an existing patient entry.

Other changes:
- create_patient_table - docstring added to improve comprehension. full_url field changed to Unique to catch out duplicates.