Function - Create Resource Table, Add Event Entry created
As there is no limit to the different resource types that can be created (the range in the sample data had between 16-20 different variants), then aside from the patient table (which needs to be created first to pass the id to others), it did not make sense to hard code any tables as this would not be maintainable.

Instead, create_resource_table has been designed to create a table of any name, so long as they have the usual primary entry keys (fullUrl, request, resource). It creates similar columns to create_patient_table, aside from the addition of a patient_id column which uses the foreign key patient.pid to ensure proper relational mapping.

Similarly, add_event_entry allows us to add any event to the input table_name, with basic error handling for duplicate values, and where the patient_id does not exist.

In preparation for the creation and insertion of event data to these dynamically-created tables, I have created another util function - pascal_to_snake_case which will aid us in changing the FHIR resource type values (currently in PascalCase), to snake_case, which is the convention used for table naming.